extension(checksum): update checksum extension to use Multihash
Instead of defining the hashing algorithm within the key, use a self-identifiable hash
([Multihash]). This way the stac spec doesn't need to introduce a long list of supported
hash algorithms, but can rely on the ones specified in [Multicodec].

I also changed the examples to use BLAKE2 instead of MD5. While MD5 is still in wide
use, I want to encourage people to update to newer hash function, which makes also
sense if you want to check the integrity of untrusted sources.

[Multihash]: https://github.com/multiformats/multihash
[Multicodec]: https://github.com/multiformats/multicodec/

Closes #503.