extension(SAR): update checksum in example
The checksum extension was update, hence also update the usage in the SAR example.

I also changed the example to use BLAKE2 instead of MD5. While MD5 is still in wide
use, I want to encourage people to update to newer hash function, which makes also
sense if you want to check the integrity of untrusted sources.