Only use "cite-as" link to refer to main publication for data.
Previous to this commit, the 'cite-as' rel type was used to describe
links to both the main publication for the dataset as well as any
publications citing the data as listed in the 'publications'
property. As discussed in #915, this can cause confusion as to how the
data should be cited. This change specifies that only the main
publication descripting the data should have a link with 'cite-as' as
the rel type in the links section.

Fixes #915