Update instrument.json to include a target.
We have been using the STAC specification for non-terrestrially focused data, e.g., the Mars. For non-Earth focused data, we need some mechanism to identify the target of the observation. By target, we mean the planet, moon, asteroid, comet, etc. that was the primary objective of the observation. If a given instrument imaged a single body over the mission lifecycle, we could make use of the mission string. For many missions, observations are captured enroute (for example a few quick snaps of Venus during a gravity assist on the way to the outer planets) or multiple targets can exist (sensors targeting Mars can be rolled to capture data of Phobos for example).

The instrument section of the item specification seems to be the most appropriate place to inject a target keyword. We envision the target keyword being used to identify the body/bodies which was the target of a particular observation. For example:

```
{"instruments":["Kaguya Terrain Camera"], 
"mission":"Kaguya/SELENE", 
"gsd": "6",
"target":"Moon"}
```